pyproject.toml
[tool.poetry]
name = "api-env"
version = "0.1.0"
description = ""
authors = ["yuenfu001 <liuyuenfu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-decouple = "^3.8"
jupyter = "^1.0.0"
pandas = "^2.1.3"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
